@@ -1,4 +1,4 @@
-__version__ = '1.2.6'
+__version__ = '1.2.7'
 
 from utilix import uconfig
 from .common import *
@@ -10,16 +10,16 @@ from .itp_map import *
 from .matplotlib_utils import *
 from .mini_analysis import *
 from .misc import *
-from .mongo_storage import *
-from .online_monitor import *
-from .rundb import *
+
 from .scada import *
 from .bokeh_utils import *
-from .rucio import *
 from .url_config import *
 
 from . import plugins
 from .plugins import *
+
+from . import storage
+from .storage import *
 
 from . import analyses
 
