@@ -46,9 +46,10 @@ def _check_pulse_count(records):
 
         # Check baseline values:
         # nan does not exist for integer:
-        mean = straxen.NO_PULSE_COUNTS if np.isnan(np.mean(rc0['baseline'])) else int(np.mean(rc0['baseline']))
-        assert count['baseline_mean'][ch] == mean
-        if not np.isnan(count['baseline_rms_mean'][ch]):
-            assert count['baseline_rms_mean'][ch] == np.mean(rc0['baseline_rms'])
-        else:
-            assert np.isnan(np.mean(rc0['baseline_rms']))
+        if len(rc0):
+            mean = straxen.NO_PULSE_COUNTS if np.isnan(np.mean(rc0['baseline'])) else int(np.mean(rc0['baseline']))
+            assert count['baseline_mean'][ch] == mean
+            if not np.isnan(count['baseline_rms_mean'][ch]):
+                assert count['baseline_rms_mean'][ch] == np.mean(rc0['baseline_rms'])
+            else:
+                assert np.isnan(np.mean(rc0['baseline_rms']))
